Specify version numbers in root pom.xml only
This removes any hard-coded version numbers from pom.xml except for the
root; the root contains a set of properties that define version numbers.

In some cases this will change the minor version used; where this
happens, the version used will always move forward, not backward; and
the major version does not change.