Fix SoftwareProcess.ServiceNotUpDiagnosticsCollector race
Previously we subscribed to serviceUp and then called onUpdated.
However, this meant that it could execute onUpdated in two threads
concurrently (if we very quickly got a callback of serviceUp changing).

Instead we now use “notifyOfInitialValue”, which will call us in our
event-handling thread with the initial value of the sensor as a 
pseudo-event for the subscription manager.