Fix VanillaSoftwareProcess install
Previously, if download and install-cmd were supplied, only the
download would be executed.

Previously if two VanillaSoftwareProcesses were used on the same
machine where neither had a downloadUrl, it would have skipped the
install of the second (sharing the same install folder).

Previously there was poor test coverage of VanillaSoftwareProcess.